+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Features
+* Added the `messaging` package which adds support for Firebase Cloud Messaging.
+
+## Changes
 * `Firebase` struct in the `database` package was renamed to `database`
 * The realtime database code has been moved from `github.com/zabawaba99/firego`
 to `github.com/zabawaba99/firego/database`